Change repro to use an array of floats and an int field in managed and use a float array for padding in native to force an SSE classification on the first byte.

+++ tests/src/Interop/StructMarshalling/PInvoke/MarshalStructAsParamDLL.cpp
@@ -1184,7 +1184,7 @@ extern "C" DLL_EXPORT BOOL STDMETHODCALLTYPE MarshalStructAsParam_AsExpByValOver
     return str.a == expected;
 }
 
-extern "C" DLL_EXPORT BOOL STDMETHODCALLTYPE MarshalStructAsParam_AsExpByValOverlappingMultipleEightByte(OverlappingMultipleEightbyte str, int i1, int i2, int i3)
+extern "C" DLL_EXPORT BOOL STDMETHODCALLTYPE MarshalStructAsParam_AsExpByValOverlappingMultipleEightByte(OverlappingMultipleEightbyte str, float i1, float i2, float i3)
 {
     return str.arr[0] == i1 && str.arr[1] == i2 && str.arr[2] == i3;
 }
